refactor: 공용 모듈, css
- logo를 common으로 이동
- css 리팩토링
