+++ src/frontend/web/src/components/Common/Logo.tsx
